+++ src/core/watchQueryOptions.ts
@@ -65,6 +65,12 @@ export interface WatchQueryOptions extends ModifiableWatchQueryOptions {
    */
   // TODO REFACTOR: rename this to document. Didn't do it yet because it's in a lot of tests.
   query: Document;
+
+  /**
+   * Arbitrary metadata stored in Redux with this query.  Designed for debugging,
+   * developer tools, etc.
+   */
+  metadata?: any;
 }
 
 // This interface is deprecated because we no longer pass around fragments separately in the core.
@@ -80,6 +86,12 @@ export interface DeprecatedWatchQueryOptions extends ModifiableWatchQueryOptions
    * referenced from the query document.
    */
   fragments?: FragmentDefinition[];
+
+  /**
+   * Arbitrary metadata stored in Redux with this query.  Designed for debugging,
+   * developer tools, etc.
+   */
+  metadata?: any;
 }
 
 export interface FetchMoreQueryOptions {